requirements.txt
langchain>=0.1.16
langchain-core>=0.1.50
langgraph>=0.0.34
langchain-openai>=0.1.6
python-dotenv
streamlit>=1.33.0
langchain-community
sentence-transformers
scikit-learn